Add kymux command line tool
A kymux command line tool is useful for debugging and testing.

To run a server listening on QUIC (UDP) port 1234, accepting one client
which will produce a video stream:

    kymux --listen 1234 video

The kymux client must connect to this server:

    kymux --connect <server_ip>:1234

The clients of the kymux nodes must connect to TCP on port 9090
(hardcoded for now).