Send the endpoint id on a stream during creation
ControlMsg::StreamOpened was used previously to notify the
Endpoint <-> Stream id. This lead to a complicated implementation that
had to receive but the ControlMsg and detect the Quic stream opening.

The endpoint id is now sent on the Quic stream during creation, instead
of sending a null byte. This allow to do the mapping directly.