router: Request to stop the tasks
This drops the references to quinn::Connection that router's tasks retain.

This allow to close the socket as soon as possible. Without this, quinn isn't
able to create a new quinn::Connection on the same port until the previous
one timeout. On timeout, all accept*() operation are returning an error and
the tasks stop.