Make the API Send Sync only on non-wasm32
The web_sys types are not Send (JsValue may not be Send because it can
only be used from the worker context).

With tokio, we need Send + Sync types to pass them to async tasks.
In Wasm, this is not a problem, since
wasm_bindgen_futures::spawn_local() accept non-Send Futures.

Therefore, conditionnally use Send types.

Refs <https://github.com/rustwasm/wasm-bindgen/issues/2833>
Refs <https://github.com/rustwasm/wasm-bindgen/issues/2409>