kymux: ipc: Decouple `kymux::Connection` and `IpcHandler`
This removes the systematic IPC implementation embedded in a
`kymux::Connection`. This caused issues while implementing the
multiclient support in kyber. The server kept returning connections
with a new IPC port attached to them.

This patch makes the IPCHandler constructible by kymux users and removes
remove its dependence from `kymux::Connection`. The user can now manage
the IPC connections however it wants.

A handy "All in one" helper is kept (`IPCForwardableConnection`), to
manage both the IPC connection and the actual kyproto implementation in
a simple and straightforward way. This will likely still be used as is
in kymux client only scenario (kyber's `libclient`).