Fix deadline reset for unreliable protocols
Once set, a deadline was never removed, causing a liveloop (it was
always active to wake up an async sleep() call from a tokio::select!
macro).

The problem was not visible with wasmtimer-rs because sleep() did not
wake up properly, so we noticed it after switching to kywasmtime. It
could in theory occur with the desktop version, but it requires certain
conditions to be triggered.