kyproto: Handle kynet stream order in GopStream
The GopStream protocol incorrectly assumed that kynet streams are
accepted (received) in order, which is not guaranteed by the underlying
QUIC protocol. In particular, it always used the latest accepted kynet
stream as the current one.

Since the underlying QUIC stream id is not exposed, add an increasing
identifier (gop_id) for all streams, and only consider data received on
the newest stream.