kycom: Split kycom logic
KyCom allows forwarding data in both directions between a TCP socket and
a kyproto endpoint/protocol.

In addition to performing data forwarding, TcpForwarder was responsible
for serializing and deserializing kyproto structs.

Extract this logic into a separate layer, so that it can be reused
independently. This will be useful for exposing an API for a KyCom
client.