kyproto: Explicitly convert foreign errors
Previously, all errors were automatically converted to ProtocolError in
protocol implementations.

Once ProtocolError is moved to a separate crate, this will violate the
Rust orphan rule, which states that a trait implementation is only
allowed if either the trait or at least one of the types in the
implementation is defined in the current crate.

This change prepares for moving ProtocolError to the kyproto-type crate.